requirements.txt
torch>=2.0.0
numpy>=1.21.0
python-chess>=1.9.0
tqdm>=4.65.0
requests>=2.31.0
zstandard>=0.21.0